Update Update/Signal/Query documentation
