+++ packages/test/src/test-integration-update.ts
@@ -257,40 +257,3 @@ test('Two Updates in first WFT', async (t) => {
     t.deepEqual(wfResult, ['1', 'done', '$']);
   });
 });
-
-/* BEGIN: Test example from WorkflowHandle docstring */
-export const incrementSignal = wf.defineSignal<[number]>('increment');
-export const getValueQuery = wf.defineQuery<number>('getValue');
-export const incrementAndGetValueUpdate = wf.defineUpdate<number, [number]>('incrementAndGetValue');
-
-export async function counterWorkflow(initialValue: number): Promise<void> {
-  let count = initialValue;
-  wf.setHandler(incrementSignal, (arg: number) => {
-    count += arg;
-  });
-  wf.setHandler(getValueQuery, () => count);
-  wf.setHandler(incrementAndGetValueUpdate, (arg: number): number => {
-    count += arg;
-    return count;
-  });
-  await wf.condition(() => false);
-}
-
-test('Update/Signal/Query example in WorkflowHandle docstrings works', async (t) => {
-  const { createWorker, startWorkflow, assertWorkflowFailedError } = helpers(t);
-  const worker = await createWorker();
-  await worker.runUntil(async () => {
-    const wfHandle = await startWorkflow(counterWorkflow, { args: [2] });
-    await wfHandle.signal(incrementSignal, 2);
-    const queryResult = await wfHandle.query(getValueQuery);
-    t.is(queryResult, 4);
-    const updateResult = await wfHandle.executeUpdate(incrementAndGetValueUpdate, { args: [2] });
-    t.is(updateResult, 6);
-    const secondUpdateHandle = await wfHandle.startUpdate(incrementAndGetValueUpdate, { args: [2] });
-    const secondUpdateResult = await secondUpdateHandle.result();
-    t.is(secondUpdateResult, 8);
-    await wfHandle.cancel();
-    await assertWorkflowFailedError(wfHandle.result(), wf.CancelledFailure);
-  });
-});
-/* END: Test example from WorkflowHandle docstring */
